Move around test main
